+++ examples/find_invariants/baseline.py
@@ -8,10 +8,10 @@ from collections.abc import Sequence
 from dataclasses import dataclass
 from typing import Never
 
-import why3_utils as why3
-
 import delphyne as dp
 from delphyne import Branch, Compute, Strategy, dfs, strategy
+
+import why3_utils as why3
 
 # fmt: off
 
@@ -46,7 +46,7 @@ def check_invariants(
     prog: why3.File, invariants: Sequence[why3.Formula]
 ) -> dp.Strategy[Compute, object, why3.File | dp.Error]:
     annotated = why3.add_invariants(prog, invariants)
-    feedback = yield from dp.compute(why3.check, prog, annotated)
+    feedback = yield from dp.compute(why3.check)(prog, annotated)
     if feedback.success:
         return annotated
     feedback.obligations = [